refactor: update app unit test, adding sinon
