+++ app/nodejs/package.json
@@ -29,6 +29,7 @@
         "got": "^14.0.0",
         "mocha": "^10.0.0",
         "nodemon": "^3.0.0",
+        "sinon": "^17.0.1",
         "supertest": "^6.1.3"
     }
 }
